==> y3_data/07_test_results.qmd <==
---
title: "Test Results"
execute:
  echo: false
  message: false
  warning: false
format: 
  html:
    embed-resources: false
editor: visual
---

```{r}
#| echo: false
#| message: false
#| warning: false

## Housekeeping

## ---------------------------
## libraries
## ---------------------------
library(tidyverse)
library(readxl)
library(openxlsx)
library(janitor)
library(kableExtra)
library(pwr)
library(kableExtra)
library(knitr)

## ---------------------------
## directory paths
## ---------------------------

#see current directory
#getwd()

#set current directory
code_file_dir<-file.path("C:/Users/yjeff/Desktop/Community Schools/Teacher Retention Data",
                "Analyses","cs_teach_rent2",  "y3_data")

data_file_dir<-file.path("C:/Users/yjeff/Box/LAUSD TR Data Year 3")

## ---------------------------
## helper functions
## ---------------------------

update_tbl_names<-function(df){
  df_update<-df
  colnames(df_update)<-c("School", "2019", "2020", "2021", "2022", "2023")
  return(df_update)
}

## ---------------------------
## load & inspect data
## ---------------------------

# load(file.path(code_file_dir, "rent_tbls.RData"))
# load(file.path(code_file_dir, "rent_plots.RData"))

#toc tables
# load(file.path(code_file_dir, "toc_rent_tbls.RData"))
# load(file.path(code_file_dir, "rent_plots_toc.RData"))

load(file.path(code_file_dir, "test_tbls.RData"))


## Add Lists
source(file.path(code_file_dir, "00_school_lists.R"))


```

## Purpose

The purpose of this document is to show the test results comparing Community Schools with their Traditional Public Schools.

## T-Tests

## Overall

**T-tests**

<details>

```{r, echo = FALSE, results='asis'}

t_test<-full_sample_test %>% mutate(
  veteran_status = "all teachers"
) %>% select(veteran_status, everything()) %>% 
  select(-c(conf_low, conf_high, statistic))
  

t_test2<-overall_t_test_df %>%
  filter(type == "overall", version == "original_5") %>% 
  select(-c(conf_low, conf_high, statistic, type, version)) #%>% 

t_test<-rbind(t_test, t_test2)

rownames(t_test)<-NULL  

colnames(t_test)<-c("Veteran Status",
                    "CS", "TS",
                    "Diff.", "p-value", "df")

t_test %>% kable()
```

</details>

**Proportions test**

<details>

::: panel-tabset
#### Overall

```{r, echo = FALSE, results='asis'}

pr_test<-prop_test[["overall"]] %>% 
  select(-c(conf_low, conf_high, statistic, parameter))

pr_test<-pr_test %>% select(year, estimate_school1, estimate_school2,
                            diff, p_value)

colnames(pr_test)<-c("Year", "CS", "TS", "Diff.",  "P-value")

pr_test %>% kable()

```

#### 0-3 Years

```{r, echo = FALSE, results='asis'}

pr_test<-prop_test[["0-3 years"]] %>% 
  select(-c(conf_low, conf_high, statistic, parameter))

colnames(pr_test)<-c("Year", "CS", "TS", "P-value")

pr_test %>% kable()

```

#### 4-5 Years

```{r, echo = FALSE, results='asis'}

pr_test<-prop_test[["4-5 years"]] %>% 
  select(-c(conf_low, conf_high, statistic, parameter))

pr_test<-pr_test %>% select(year, estimate_school1, estimate_school2,
                            diff, p_value)

colnames(pr_test)<-c("Year", "CS", "TS", "Diff.",  "P-value")

pr_test %>% kable()

```

#### 6-10 Years

```{r, echo = FALSE, results='asis'}

pr_test<-prop_test[["6-10 years"]] %>% 
  select(-c(conf_low, conf_high, statistic, parameter))

pr_test<-pr_test %>% select(year, estimate_school1, estimate_school2,
                            diff, p_value)

colnames(pr_test)<-c("Year", "CS", "TS", "Diff.",  "P-value")

pr_test %>% kable()

```

#### 11-15 Years

```{r, echo = FALSE, results='asis'}

pr_test<-prop_test[["11-15 years"]] %>% 
  select(-c(conf_low, conf_high, statistic, parameter))

pr_test<-pr_test %>% select(year, estimate_school1, estimate_school2,
                            diff, p_value)

colnames(pr_test)<-c("Year", "CS", "TS", "Diff.",  "P-value")

pr_test %>% kable()

```

#### 15+ Years

```{r, echo = FALSE, results='asis'}

pr_test<-prop_test[["15+ years"]] %>% 
  select(-c(conf_low, conf_high, statistic, parameter))

pr_test<-pr_test %>% select(year, estimate_school1, estimate_school2,
                            diff, p_value)

colnames(pr_test)<-c("Year", "CS", "TS", "Diff.",  "P-value")

pr_test %>% kable()

```
:::

</details>

## By Neighborhood

Notes: NA values are due to missing cases. For example, if a Community School does not have any 4-5 teachers for a given year, then running t-test results was not available.

### All Teachers

<details>

::: panel-tabset
#### Full

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test %>% kable()
```

#### Curated

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test<-t_test %>% 
  filter(`p-value`< 0.05) 

t_test %>% kable()
```
:::

</details>

### 0-3 Years

<details>

::: panel-tabset
#### Full

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test_veteran[["original_5"]][["0-3 years"]] %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test %>% kable()

```

#### Curated

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test_veteran[["original_5"]][["0-3 years"]] %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test<-t_test %>% 
  filter(`p-value`< 0.05) 

t_test %>% kable()
```
:::

</details>

### 4-5 Years

<details>

::: panel-tabset
#### All

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test_veteran[["original_5"]][["4-5 years"]] %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test %>% kable()

```

#### Curated

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test_veteran[["original_5"]][["4-5 years"]] %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test<-t_test %>% 
  filter(`p-value`< 0.05) 

t_test %>% kable()
```
:::

</details>

### 6-10 Years

<details>

::: panel-tabset
#### All

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test_veteran[["original_5"]][["6-10 years"]] %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test %>% kable()

```

#### Curated

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test_veteran[["original_5"]][["6-10 years"]] %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test<-t_test %>% 
  filter(`p-value`< 0.05) 

t_test %>% kable()
```
:::

</details>

### 11-15 Years

<details>

::: panel-tabset
#### Full

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test_veteran[["original_5"]][["11-15 years"]] %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test %>% kable()

```

#### Curated

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test_veteran[["original_5"]][["11-15 years"]] %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test<-t_test %>% 
  filter(`p-value`< 0.05) 

t_test %>% kable()
```
:::

</details>

### 15+ Years

<details>

::: panel-tabset
#### Full

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test_veteran[["original_5"]][["15+ years"]] %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test %>% kable()

```

#### Curated

```{r, echo = FALSE, results='asis'}

t_test<-combined_t_test_veteran[["original_5"]][["15+ years"]] %>% 
  select(-c(conf_low, conf_high, statistic)) 

colnames(t_test)<-c('Neighborhood',"CS", "TS",
                    "Diff.", "p-value", "df")

t_test<-t_test %>% 
  filter(`p-value`< 0.05) 

t_test %>% kable()
```
:::

</details>
