==> y3_data/08_cs_only_schools.qmd <==
---
title: "CS Only School Report"
execute:
  echo: false
  message: false
  warning: false
format: 
  html:
    embed-resources: false
editor: visual
---

## Purpose

1.  Analyzes retention only for the Community Schools:
    -   0-3 Years
    -   4-5 Years
    -   6-10 Years
    -   11-15 Years
    -   Over 15 Years

```{r}
#| echo: false
#| message: false
#| warning: false

## Housekeeping

## ---------------------------
## libraries
## ---------------------------
library(tidyverse)
library(readxl)
library(openxlsx)
library(janitor)
library(kableExtra)
library(pwr)
library(kableExtra)
library(knitr)

## ---------------------------
## directory paths
## ---------------------------

#see current directory
#getwd()

#set current directory
code_file_dir<-file.path("C:/Users/yjeff/Desktop/Community Schools/Teacher Retention Data","Analyses", "cs_teach_rent2","y3_data")

data_file_dir<-file.path("C:/Users/yjeff/Box/LAUSD TR Data Year 3")

## ---------------------------
## helper functions
## ---------------------------

update_tbl_names<-function(df){
  df_update<-df
  colnames(df_update)<-c("School", "2019", "2020", "2021", "2022", "2023")
  return(df_update)
}

## ---------------------------
## load & inspect data
## ---------------------------

load(file.path(code_file_dir, "veteran_rent_tbls.RData"))
load(file.path(code_file_dir, "veteran_plots.RData"))

## Add Lists
source(file.path(code_file_dir, "00_school_lists.R"))


```

# Descriptive Analyses - Retention Rates

## All Schools

<details>

### Overall

<details>

::: panel-tabset
#### Only 2022-23 Years

```{r}
cs_only_overall_plots[["22-23"]][["overall"]][["overall"]]
```

#### All Years

```{r}
cs_only_overall_plots[["all years"]][["overall"]][["overall"]]

```
:::

</details>

```{r, echo = FALSE, results='asis'}
teach_yr_string<-names(toc_rent_overall[["ts"]][["original_5"]])

#cs_plot<-cs_overall_plots[["overall"]][["original_5"]][["0-3 years"]]

gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  type<-"overall"
  df_cs<-toc_rent_overall2[["cs"]][["original_5"]][[teach_year]]
  # cs_plot<-cs_overall_plots[["overall"]][["original_5"]][[teach_yr_string[[x]]]]
 
  #neighborhood<-names(rent_by_neighborhood)[[x]]

  knit_child(
    '_07.1_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
  
})

cat(unlist(gxs), sep = '\n')

```

</details>

## Elementary Schools

<details>

### Overall

<details>

::: panel-tabset
#### Only 2022-23 Years

```{r}
cs_only_overall_plots[["22-23"]][["elem"]][["overall"]]
```

#### All Years

```{r}
cs_only_overall_plots[["all years"]][["elem"]][["overall"]]

```
:::

</details>

```{r, echo = FALSE, results='asis'}
teach_yr_string<-names(toc_rent_overall[["ts"]][["original_5"]])


gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  type<-"elem"
  df_cs<-toc_rent_overall_elem2[["cs"]][["original_5"]][[teach_year]]
  
  
  #neighborhood<-names(rent_by_neighborhood)[[x]]

  knit_child(
    '_07.1_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
  
})

cat(unlist(gxs), sep = '\n')

```

</details>

## Middle/High Schools

<details>

### Overall

<details>

::: panel-tabset
#### Only 2022-23 Years

```{r}
cs_only_overall_plots[["22-23"]][["ms_hs"]][["overall"]]
```

#### All Years

```{r}
cs_only_overall_plots[["all years"]][["ms_hs"]][["overall"]]

```
:::

</details>

```{r, echo = FALSE, results='asis'}
teach_yr_string<-names(toc_rent_overall[["ts"]][["original_5"]])


gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  type<-"ms_hs"
  df_cs<-toc_rent_overall_ms_hs2[["cs"]][["original_5"]][[teach_year]]
  
  
  #neighborhood<-names(rent_by_neighborhood)[[x]]

  knit_child(
    '_07.1_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
  
})

cat(unlist(gxs), sep = '\n')

```

</details>

## Neighborhoods

<details>

```{r, echo = FALSE, results='asis'}
neigh_string<-names(cs_only_plots[["original_5"]][["0-3 years"]])


gxs<-lapply(1:length(neigh_string), function(x){
  
  neigh <- neigh_string[[x]]
  
  neigh_name<-str_replace_all(neigh, "_", " ") #%>% str_to_title()
  df_cs_03<-cs_only_plots[["original_5"]][["0-3 years"]][[neigh]]
  df_cs_45<-cs_only_plots[["original_5"]][["4-5 years"]][[neigh]]
  df_cs_6_10<-cs_only_plots[["original_5"]][["6-10 years"]][[neigh]]
  df_cs_11_15<-cs_only_plots[["original_5"]][["11-15 years"]][[neigh]]
  df_cs_15_over<-cs_only_plots[["original_5"]][["15+ years"]][[neigh]]
  
  #neighborhood 
  df_cs_03_22_23<-cs_only_plots_22_23_neigh[["original_5"]][["0-3 years"]][[neigh]]
  df_cs_45_22_23<-cs_only_plots_22_23_neigh[["original_5"]][["4-5 years"]][[neigh]]
  df_cs_6_10_22_23<-cs_only_plots_22_23_neigh[["original_5"]][["6-10 years"]][[neigh]]
  df_cs_11_15_22_23<-cs_only_plots_22_23_neigh[["original_5"]][["11-15 years"]][[neigh]]
  df_cs_15_over_22_23<-cs_only_plots_22_23_neigh[["original_5"]][["15+ years"]][[neigh]]
  
  #overall plots (all years)
  df_cs_neigh_overall<-cs_overall_all_yrs_neigh_plots[[neigh]]
  
  #overall plots (22-23)
  df_cs_neigh_overall_22_23<-cs_overall_22_23_neigh_plots[[neigh]]
  
  #neighborhood<-names(rent_by_neighborhood)[[x]]

  knit_child(
    '_07.2_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
  
})

cat(unlist(gxs), sep = '\n')

```

</details>
