==> y3_data/09_cs_only_schools_by_cohort.qmd <==
---
title: "CS Only School Report - By Cohort"
execute:
  echo: false
  message: false
  warning: false
format: 
  html:
    embed-resources: false
editor: visual
---

## Purpose

This is an update version of the CS Only School report. There are two updates:

1.  Retention rates are displayed exclusively for the period after schools became community schools, with data from years prior to this status excluded.

2.  Report also includes retention rates broken down by cohort.

```{r}
#| echo: false
#| message: false
#| warning: false

## Housekeeping

## ---------------------------
## libraries
## ---------------------------
library(tidyverse)
library(readxl)
library(openxlsx)
library(janitor)
library(kableExtra)
library(pwr)
library(kableExtra)
library(knitr)

## ---------------------------
## directory paths
## ---------------------------

#see current directory
#getwd()

#set current directory
code_file_dir<-file.path("C:/Users/yjeff/Desktop/Community Schools/Teacher Retention Data","Analyses", "cs_teach_rent2","y3_data")

data_file_dir<-file.path("C:/Users/yjeff/Box/LAUSD TR Data Year 3")

## ---------------------------
## helper functions
## ---------------------------

update_tbl_names<-function(df){
  df_update<-df
  colnames(df_update)<-c("School", "2019", "2020", "2021", "2022", "2023")
  return(df_update)
}

## ---------------------------
## load & inspect data
## ---------------------------

load(file.path(code_file_dir, "combined_cohort_rent_tbls.RData"))
load(file.path(code_file_dir, "by_cohort_plots.RData"))

## Add Lists
source(file.path(code_file_dir, "00_school_lists.R"))
```

# Descriptive Analyses - Retention Rates

## Combined Overall

Includes retention rates only from the time schools became community schools, excluding data from years prior to their community school status.

<details>

```{r, echo = FALSE, results='asis'}
teach_yr_string<-names(combined_overall_bp[["22_23"]])

gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  plot_list_22_23<-combined_overall_bp[["22_23"]]
  plot_list_all_years<-combined_overall_bp[["all_yrs"]]
  rent_tbl<-combined_overall_rent_tbls

  knit_child(
    '_07.3_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

</details>

## By Cohort Analysis

### Cohort 1

<details>

```{r, echo = FALSE, results='asis'}
#overall Cohort
teach_yr_string<-c("overall")

gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  plot_list_22_23<-overall_bp[["22_23"]][["cohort_1"]]
  plot_list_all_years<-overall_bp[["all_yrs"]][["cohort_1"]]
  rent_tbl<-cohort_overall_v2[["cohort_1"]]
  
  knit_child(
    '_07.4_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')
```

```{r, echo = FALSE, results='asis'}
teach_yr_string<-names(cohort_bp[["22_23"]][["cohort_1"]])

gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  plot_list_22_23<-cohort_bp[["22_23"]][["cohort_1"]]
  plot_list_all_years<-cohort_bp[["all_yrs"]][["cohort_1"]]
  rent_tbl<-cohort_overall_teach_exp_v2[["cohort_1"]]    
  

  knit_child(
    '_07.3_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')
```

</details>

### Cohort 2

<details>

```{r, echo = FALSE, results='asis'}
#overall Cohort
teach_yr_string<-c("overall")

gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  plot_list_22_23<-overall_bp[["22_23"]][["cohort_2"]]
  plot_list_all_years<-overall_bp[["all_yrs"]][["cohort_2"]]
  rent_tbl<-cohort_overall_v2[["cohort_2"]]
  
  knit_child(
    '_07.4_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')
```

```{r, echo = FALSE, results='asis'}
teach_yr_string<-names(cohort_bp[["22_23"]][["cohort_2"]])

gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  plot_list_22_23<-cohort_bp[["22_23"]][["cohort_2"]]
  plot_list_all_years<-cohort_bp[["all_yrs"]][["cohort_2"]]
  rent_tbl<-cohort_overall_teach_exp_v2[["cohort_2"]]

  knit_child(
    '_07.3_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')
```

</details>

### Cohort 3

<details>

```{r, echo = FALSE, results='asis'}
#overall Cohort
teach_yr_string<-c("overall")

gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  plot_list_22_23<-overall_bp[["22_23"]][["cohort_3"]]
  plot_list_all_years<-overall_bp[["all_yrs"]][["cohort_3"]]
  rent_tbl<-cohort_overall_v2[["cohort_3"]]
  
  knit_child(
    '_07.4_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')
```

```{r, echo = FALSE, results='asis'}
teach_yr_string<-names(cohort_bp[["22_23"]][["cohort_3"]])

gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  plot_list_22_23<-cohort_bp[["22_23"]][["cohort_3"]]
  plot_list_all_years<-cohort_bp[["all_yrs"]][["cohort_3"]]
  rent_tbl<-cohort_overall_teach_exp_v2[["cohort_3"]]

  knit_child(
    '_07.3_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')
```

</details>

## Cohort by Cohort Comparison

### Interpretation

<details>

Among teachers with 0–3 years of experience, retention patterns vary by cohort. Cohort 1 started with retention rates in the 70s, rising sharply to 95% in 2023. Cohort 2 began in the 80s but declined to the mid-70s by 2022–2023. Cohort 3 dropped from 90% in 2021 to 70% by 2023. Similar trends are seen among teachers with 4–5 years of experience.

For 6–10 year veterans, Cohort 1’s retention rates declined steadily from 2020 to 2023. Cohort 2 began at 100% but fell to the mid-80s to low-90s range. Cohort 3 rose from 58% in 2021 to 92% in 2022, then declined to 76% in 2023.

Among 11–15 year veterans, Cohort 1 maintained retention in the 80s throughout. Cohort 2 increased from 72% to 93%, while Cohort 3 dropped from 100% in 2021 to 75% in 2022 before rebounding to 90% in 2023.

For those with 15+ years, Cohorts 1 and 2 maintained steady retention in the high-80s and mid-80s, respectively. Cohort 3 saw a decline in retention from 2021 to 2023.

</details>

```{r, echo = FALSE, results='asis'}
teach_yr_string<-c("Overall",
                   names(cohort_bp[["22_23"]][["cohort_3"]]))

gxs<-lapply(1:length(teach_yr_string), function(x){
  
  teach_year <- teach_yr_string[[x]]
  
  if (x == 1){
    cohort_1_plot<-
      overall_bp[["all_yrs"]][["cohort_1"]]
    cohort_2_plot<-
      overall_bp[["all_yrs"]][["cohort_2"]]
    cohort_3_plot<-
      overall_bp[["all_yrs"]][["cohort_3"]]
  }
  
  if (x >1){
    cohort_1_plot<-cohort_bp[["all_yrs"]][["cohort_1"]][[teach_year]]
    cohort_2_plot<-cohort_bp[["all_yrs"]][["cohort_2"]][[teach_year]]
    cohort_3_plot<-cohort_bp[["all_yrs"]][["cohort_3"]][[teach_year]]
  }
  
  knit_child(
    '_07.5_report_full - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')
```
