==> y3_data/11_veteran_demographics.qmd <==
---
title: "Veteran Teacher Report - Demographics"
execute:
  echo: false
  message: false
  warning: false
format: 
  html:
    embed-resources: false
editor: visual
---

## Purpose

Provide demographics by school veteran status.

```{r}
#| echo: false
#| message: false
#| warning: false

## Housekeeping

## ---------------------------
## libraries
## ---------------------------
library(tidyverse)
library(kableExtra)
library(pwr)
library(knitr)

## ---------------------------
## directory paths
## ---------------------------

#see current directory
#getwd()

#set current directory
code_file_dir<-file.path("C:/Users/yjeff/Desktop/Community Schools/Teacher Retention Data","Analyses", "cs_teach_rent2","y3_data")

data_file_dir<-file.path("C:/Users/yjeff/Box/LAUSD TR Data Year 3")

## ---------------------------
## helper functions & strings
## ---------------------------

school_name_string<-c("all", "Elementary", "Middle/High School")

## ---------------------------
## load & inspect data
## ---------------------------

load(file.path(code_file_dir, "vet_combined_tbls.RData"))
load(file.path(code_file_dir, "demo_combined_tbls.RData"))
load(file.path(code_file_dir,'lausd_veteran.RData'))
load(file.path(code_file_dir,"veteran_percent_plots.RData"))
load(file.path(code_file_dir, "lausd_vet_data.RData"))

## Add Lists
#source(file.path(code_file_dir, "00_school_lists.R"))

```

# LAUSD Data

## LAUSD HR Data

<details>

```{r, echo = FALSE, results='asis'}

gxs<-lapply(1:1, function(x){
  
  knit_child(
    '_lausd_data - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

</details>

## State Data

<details>

::: panel-tabset
#### Count

```{r}
veteran_status_wide_list[["count"]] %>% kable()
```

#### Percent

```{r}
veteran_status_wide_list[["percent"]] %>% kable()
```
:::

</details>

</details>

# BIPOC Demographic Data

## Derived BIPOC Veteran Teachers

<details>

::: panel-tabset
#### 2020-2021

```{r}
perc_tbl_bipoc[["2020_21"]] %>% kable()
```

In **2020-2021:**

Among **Elementary Teachers:**

-   `r perc_tbl_bipoc[["2020_21"]][1,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2020_21"]][1,12]`% are **BIPOC**

Among **Secondary Teachers:**

-   `r perc_tbl_bipoc[["2020_21"]][2,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2020_21"]][2,12]`% are **BIPOC**

Among **Special Education Teachers:**

-   `r perc_tbl_bipoc[["2020_21"]][3,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2020_21"]][3,12]`% are **BIPOC**

Among **Teachers Overall:**

-   `r perc_tbl_bipoc[["2020_21"]][4,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2020_21"]][4,12]`% are **BIPOC**

#### 2021-2022

```{r}
perc_tbl_bipoc[["2021_22"]] %>% kable()

```

In **2021-2022:**

Among **Elementary Teachers:**

-   `r perc_tbl_bipoc[["2021_22"]][1,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2021_22"]][1,12]`% are **BIPOC**

Among **Secondary Teachers:**

-   `r perc_tbl_bipoc[["2021_22"]][2,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2021_22"]][2,12]`% are **BIPOC**

Among **Special Education Teachers:**

-   `r perc_tbl_bipoc[["2021_22"]][3,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2021_22"]][3,12]`% are **BIPOC**

Among **Teachers Overall:**

-   `r perc_tbl_bipoc[["2021_22"]][4,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2021_22"]][4,12]`% are **BIPOC**

#### 2022-2023

```{r}
perc_tbl_bipoc[["2022_23"]] %>% kable()
```

In **2022-2023:**

Among **Elementary Teachers:**

-   `r perc_tbl_bipoc[["2022_23"]][1,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2022_23"]][1,12]`% are **BIPOC**

Among **Secondary Teachers:**

-   `r perc_tbl_bipoc[["2022_23"]][2,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2022_23"]][2,12]`% are **BIPOC**

Among **Special Education Teachers:**

-   `r perc_tbl_bipoc[["2022_23"]][3,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2022_23"]][3,12]`% are **BIPOC**

Among **Teachers Overall:**

-   `r perc_tbl_bipoc[["2022_23"]][4,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2022_23"]][4,12]`% are **BIPOC**

#### 2023-2024

```{r}
perc_tbl_bipoc[["2023_24"]] %>% kable()
```

In **2023-2024:**

Among **Elementary Teachers:**

-   `r perc_tbl_bipoc[["2023_24"]][1,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2023_24"]][1,12]`% are **BIPOC**

Among **Secondary Teachers:**

-   `r perc_tbl_bipoc[["2023_24"]][2,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2023_24"]][2,12]`% are **BIPOC**

Among **Special Education Teachers:**

-   `r perc_tbl_bipoc[["2023_24"]][3,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2023_24"]][3,12]`% are **BIPOC**

Among **Teachers Overall:**

-   `r perc_tbl_bipoc[["2023_24"]][4,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2023_24"]][4,12]`% are **BIPOC**

#### 2024-2025

```{r}
perc_tbl_bipoc[["2024_25"]] %>% kable()
```

In **2024-2025:**

Among **Elementary Teachers:**

-   `r perc_tbl_bipoc[["2024_25"]][1,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2024_25"]][1,12]`% are **BIPOC**

Among **Secondary Teachers:**

-   `r perc_tbl_bipoc[["2024_25"]][2,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2024_25"]][2,12]`% are **BIPOC**

Among **Special Education Teachers:**

-   `r perc_tbl_bipoc[["2024_25"]][3,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2024_25"]][3,12]`% are **BIPOC**

Among **Teachers Overall:**

-   `r perc_tbl_bipoc[["2024_25"]][4,11]`% are **BIPOC & have 10+ years of teaching experience**, `r perc_tbl_bipoc[["2024_25"]][4,12]`% are **BIPOC**
:::

</details>

## Process

<details>

### Formulas

<details>

We used the following formulas to calculate the BIPOC values:

**Calculate the Number of BIPOC Veteran Teachers** $$ 
\text{(Total Number of Veteran Teachers)} \times \text{(Proportion BIPOC)}
$$ $$
= \text{Number of BIPOC Veteran Teachers} 
$$

For example, below is a more precise formula to calculate the **number of BIPOC teachers with 10+ years of experience**.

$$ 
\begin{aligned}
\text{(Total Number of Teachers with over 10+ years of experience)} 
\times \text{(Proportion BIPOC)} \\ 
= \text{Number of BIPOC Teachers with over 10 years of experience} 
\end{aligned}
$$

**Calculate Percent BIPOC Veteran Teachers**

$$
\frac{\text {Number of BIPOC Veteran Teachers}} {\text {Number of All Teachers}} = 
\text{Percent of BIPOC Veteran Teachers}
$$

</details>

### Number of BIPOC Teachers by Veteran Status

To determine the number of BIPOC teachers by veteran status, we first determined the percentage of BIPOC teachers in the sample (see tables below).

::: panel-tabset
#### 2020-2021

```{r}
bipoc_calculated_lausd_data[["2020_21"]] %>% kable()
```

#### 2021-2022

```{r}
bipoc_calculated_lausd_data[["2021_22"]] %>% kable()
```

#### 2022-2023

```{r}
bipoc_calculated_lausd_data[["2022_23"]] %>% kable()
```

#### 2023-2024

```{r}
bipoc_calculated_lausd_data[["2023_24"]] %>% kable()
```

#### 2024-2025

```{r}
bipoc_calculated_lausd_data[["2024_25"]] %>% kable()
```
:::

Using the first formula, we then calculated the number of BIPOC teachers by veteran status.

::: panel-tabset
#### 2020-2021

```{r}
bipoc_count_tbl[["tv_2020_21"]] %>% kable()
```

#### 2021-2022

```{r}
bipoc_count_tbl[["tv_2021_22"]] %>% kable()
```

#### 2022-2023

```{r}
bipoc_count_tbl[["tv_2022_23"]]%>% kable()
```

#### 2023-2024

```{r}
bipoc_count_tbl[["tv_2023_24"]]%>% kable()
```

#### 2024-2025

```{r}
bipoc_count_tbl[["tv_2024_25"]]%>% kable()
```
:::

With this table, we then use the second formula to determine the percentage of BIPOC veteran teachers (see **Derived BIPOC Veteran Teachers** section).

</details>

</details>

# Combined Version

# 5 Categories

Includes these categories:

-   0-3 years
-   4-5 years
-   6-10 years
-   11-15 years
-   15+ years

<details>

## All Teachers

<details>

```{r, echo = FALSE, results='asis'}
school_type_string<-names(vet_combined_tbls[["version1"]][["overall"]][["cs"]])

gxs<-lapply(1:length(school_type_string), function(x){
  
  school_type <- school_type_string[[x]]
  school_name <- school_name_string[[x]]
  
  count_tbl<-vet_combined_tbls[["version1"]][["overall"]][["cs"]][[school_type]][["teachers"]][["all"]][["count"]]
  
  percent_tbl<-vet_combined_tbls[["version1"]][["overall"]][["cs"]][[school_type]][["teachers"]][["all"]][["percent"]]

  graphs<-plots[["version1"]][[school_type]][["teachers"]][["all"]]
  
  knit_child(
    '_11.1_veteran_demographics - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

## Teachers of color

<details>

```{r, echo = FALSE, results='asis'}
school_type_string<-names(vet_combined_tbls[["version1"]][["overall"]][["cs"]])

gxs<-lapply(1:length(school_type_string), function(x){
  
  school_type <- school_type_string[[x]]
  school_name <- school_name_string[[x]]
  
  count_tbl<-vet_combined_tbls[["version1"]][["overall"]][["cs"]][[school_type]][["teachers"]][["color"]][["count"]]
  
  percent_tbl<-vet_combined_tbls[["version1"]][["overall"]][["cs"]][[school_type]][["teachers"]][["color"]][["percent"]]

  graphs<-plots[["version1"]][[school_type]][["teachers"]][["color"]]
  
  knit_child(
    '_11.1_veteran_demographics - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

</details>

</details>

# 2 Categories

Includes these categories:

-   0-5 years
-   5+ years

<details>

## All Teachers

<details>

```{r, echo = FALSE, results='asis'}
school_type_string<-names(vet_combined_tbls[["version1"]][["overall"]][["cs"]])

gxs<-lapply(1:length(school_type_string), function(x){
  
  school_type <- school_type_string[[x]]
  school_name <- school_name_string[[x]]
  
  count_tbl<-vet_combined_tbls[["version2"]][["overall"]][["cs"]][[school_type]][["teachers"]][["all"]][["count"]]
  
  percent_tbl<-vet_combined_tbls[["version2"]][["overall"]][["cs"]][[school_type]][["teachers"]][["all"]][["percent"]]

  graphs<-plots[["version2"]][[school_type]][["teachers"]][["all"]]
  
  knit_child(
    '_11.2_veteran_demographics - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

## Teachers of Color

<details>

```{r, echo = FALSE, results='asis'}
school_type_string<-names(vet_combined_tbls[["version1"]][["overall"]][["cs"]])

gxs<-lapply(1:length(school_type_string), function(x){
  
  school_type <- school_type_string[[x]]
  school_name <- school_name_string[[x]]
  
  count_tbl<-vet_combined_tbls[["version2"]][["overall"]][["cs"]][[school_type]][["teachers"]][["color"]][["count"]]
  
  percent_tbl<-vet_combined_tbls[["version2"]][["overall"]][["cs"]][[school_type]][["teachers"]][["color"]][["percent"]]

  graphs<-plots[["version2"]][[school_type]][["teachers"]][["color"]]
  
  knit_child(
    '_11.2_veteran_demographics - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

</details>

</details>

# 3 Categories (Version 1)

NOTE: Teachers include In-Classroom & Out-of-Classroom Teachers

Includes these categories:

-   0-3 years
-   4-10 years
-   10+ years

<details>

## All Teachers

<details>

```{r, echo = FALSE, results='asis'}
school_type_string<-names(vet_combined_tbls[["version3"]][["overall"]][["cs"]])

gxs<-lapply(1:length(school_type_string), function(x){
  
  school_type <- school_type_string[[x]]
  school_name <- school_name_string[[x]]
  
  count_tbl<-vet_combined_tbls[["version3"]][["overall"]][["cs"]][[school_type]][["teachers"]][["all"]][["count"]]
  
  percent_tbl<-vet_combined_tbls[["version3"]][["overall"]][["cs"]][[school_type]][["teachers"]][["all"]][["percent"]]

  graphs<-plots[["version3"]][[school_type]][["teachers"]][["all"]]
  
  knit_child(
    '_11.3_veteran_demographics - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

## Teachers of Color

<details>

```{r, echo = FALSE, results='asis'}
school_type_string<-names(vet_combined_tbls[["version1"]][["overall"]][["cs"]])

gxs<-lapply(1:length(school_type_string), function(x){
  
  school_type <- school_type_string[[x]]
  school_name <- school_name_string[[x]]
  
  count_tbl<-vet_combined_tbls[["version3"]][["overall"]][["cs"]][[school_type]][["teachers"]][["color"]][["count"]]
  
  percent_tbl<-vet_combined_tbls[["version3"]][["overall"]][["cs"]][[school_type]][["teachers"]][["color"]][["percent"]]

  graphs<-plots[["version3"]][[school_type]][["teachers"]][["color"]]
  
  knit_child(
    '_11.3_veteran_demographics - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

</details>

</details>

# 3 Categories (Version 2)

NOTE: Only includes In-Classroom Teachers

Includes these categories:

-   0-3 years
-   4-10 years
-   10+ years

<details>

## All Teachers

<details>

```{r, echo = FALSE, results='asis'}
school_type_string<-names(vet_combined_tbls[["version4"]][["overall"]][["cs"]])

gxs<-lapply(1:length(school_type_string), function(x){
  
  school_type <- school_type_string[[x]]
  school_name <- school_name_string[[x]]
  
  count_tbl<-vet_combined_tbls[["version4"]][["overall"]][["cs"]][[school_type]][["teachers"]][["all"]][["count"]]
  
  percent_tbl<-vet_combined_tbls[["version4"]][["overall"]][["cs"]][[school_type]][["teachers"]][["all"]][["percent"]]

  graphs<-plots[["version4"]][[school_type]][["teachers"]][["all"]]
  
  knit_child(
    '_11.3_veteran_demographics - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

## Teachers of Color

<details>

```{r, echo = FALSE, results='asis'}
school_type_string<-names(vet_combined_tbls[["version4"]][["overall"]][["cs"]])

gxs<-lapply(1:length(school_type_string), function(x){
  
  school_type <- school_type_string[[x]]
  school_name <- school_name_string[[x]]
  
  count_tbl<-vet_combined_tbls[["version4"]][["overall"]][["cs"]][[school_type]][["teachers"]][["color"]][["count"]]
  
  percent_tbl<-vet_combined_tbls[["version4"]][["overall"]][["cs"]][[school_type]][["teachers"]][["color"]][["percent"]]

  graphs<-plots[["version4"]][[school_type]][["teachers"]][["color"]]
  
  knit_child(
    '_11.3_veteran_demographics - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

</details>

</details>

# Demographics - Percentage of Teachers of Color

<details>

```{r, echo = FALSE, results='asis'}
school_type_string<-names(demo_combined_tbls[["overall"]][["cs"]])

gxs<-lapply(1:length(school_type_string), function(x){
  
  school_type <- school_type_string[[x]]
  school_name <- school_name_string[[x]]
  
  count_tbl<-demo_combined_tbls[["overall"]][["cs"]][[school_type]][["teachers"]][["all"]][["count"]]
  
  percent_tbl<-demo_combined_tbls[["overall"]][["cs"]][[school_type]][["teachers"]][["all"]][["percent"]]

  graphs<-plots_demo[[school_type]][["teachers"]][["all"]]
  
  knit_child(
    '_11.3_veteran_demographics - child.Rmd', envir = environment(), quiet = TRUE
    
  )
})

cat(unlist(gxs), sep = '\n')

```

</details>
